+++ src/queries/utils/buildQuery.js
@@ -9,7 +9,7 @@ const buildQuery = ({
   searchIndex,
   isExcludedSeeFromLimiter = false,
 }) => {
-  const indexData = searchableIndexesMap[searchIndex || searchableIndexesValues.KEYWORD];
+  const indexData = searchableIndexesMap[searchIndex];
 
   if (!indexData) {
     return '';
@@ -33,6 +33,10 @@ const buildQuery = ({
     }
 
     if (isExcludedSeeFromLimiter) {
+      if (searchIndex === searchableIndexesValues.KEYWORD) {
+        return [`${searchableIndexesValues.KEYWORD}=="%{query}" and authRefType == "Authorized"`];
+      }
+
       return queryParts;
     }